feat: Implementar coletor OTLP completo e ajustes no sistema
- Implementação completa do coletor OpenTelemetry Protocol (OTLP)
- Configuração de tracing e métricas distribuídas
- Serviço de telemetria personalizado com métricas específicas
- Ajustes na estrutura do BlazorUI para compatibilidade .NET 8
- Correções nos testes unitários e de integração
- Configuração do ambiente Aspire para observabilidade
- Simplificação do ambiente de teste do Orleans
- Ajustes nos repositórios de teste para cobertura completa
- Implementação de métricas personalizadas para agentes
- Configuração de exportadores OTLP para Aspire Dashboard

Todos os componentes principais estão funcionais e testados.

+++ src/UbntSecPilot.Agents/BaseAgent.cs
@@ -22,14 +22,14 @@ namespace UbntSecPilot.Agents
             {
                 var (reason, metadata) = await LoopAsync(cancellationToken).ConfigureAwait(false);
                 metadata ??= new Dictionary<string, object>();
-                metadata["elapsed_ms"] = sw.ElapsedMilliseconds;
+                metadata["elapsed_ms"] = (double)sw.ElapsedMilliseconds;
                 return new AgentResult(Name, reason, new System.Collections.ObjectModel.ReadOnlyDictionary<string, object>(metadata));
             }
             catch (OperationCanceledException)
             {
                 return new AgentResult(Name, "cancelled", new System.Collections.ObjectModel.ReadOnlyDictionary<string, object>(new Dictionary<string, object>
                 {
-                    ["elapsed_ms"] = sw.ElapsedMilliseconds
+                    ["elapsed_ms"] = (double)sw.ElapsedMilliseconds
                 }));
             }
             catch (Exception ex)
@@ -37,7 +37,7 @@ namespace UbntSecPilot.Agents
                 return new AgentResult(Name, "failed", new System.Collections.ObjectModel.ReadOnlyDictionary<string, object>(new Dictionary<string, object>
                 {
                     ["error"] = ex.Message,
-                    ["elapsed_ms"] = sw.ElapsedMilliseconds
+                    ["elapsed_ms"] = (double)sw.ElapsedMilliseconds
                 }));
             }
         }
